requirements.txt
PyQt5>=5.15.0
py2app>=0.28.0; sys_platform == 'darwin'
jaraco.text>=4.0.0
jaraco.functools>=4.1.0
jaraco.context>=6.0.1
more-itertools>=10.7.0
autocommand>=2.2.2
sip>=6.0.0
pyinstaller>=6.13.0; sys_platform == 'darwin'
pillow>=11.0.0
PyYAML>=6.0.0 